fixed delete logic in cosmetics, fixed expenses delete

+++ resources/js/Components/ProcurementsComponent.jsx
@@ -55,27 +55,43 @@ const ProcurementsComponent = ({auth, cosmetics}) => {
     };
 
     const deleteProcurement = async (procurementId) => {
+
         try {
-            const response = await fetch(`/cosmetics_procurements/${procurementId}`, {
-                method: 'DELETE',
-            });
-            const responseData = await response.json(); 
 
-            if (!response.ok) {
-                Swal.fire({
-                    icon: "error",
-                    title: "Oops...",
-                    text: `Greška: ${responseData.message}`,
-                });
-            } else {
-                Swal.fire({
-                    title: "Uspješno!",
-                    text: responseData.message,
-                    icon: "success"
-                });
+            const del = Swal.fire({
+                title: 'Da li ste sigurni?',
+                text: "Ova radnja se ne može poništiti!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Da, obriši!',
+                cancelButtonText: 'Otkaži'
+            })
 
-                // Remove the deleted procurement from state
-                setProcurements(prevProcurements => prevProcurements.filter(procurement => procurement.id !== procurementId));
+            if((await del).isConfirmed) {
+
+                const response = await fetch(`/cosmetics_procurements/${procurementId}`, {
+                    method: 'DELETE',
+                });
+                const responseData = await response.json();
+
+                if (!response.ok) {
+                    Swal.fire({
+                        icon: "error",
+                        title: "Oops...",
+                        text: `Greška: ${responseData.message}`,
+                    });
+                } else {
+                    Swal.fire({
+                        title: "Uspješno!",
+                        text: responseData.message,
+                        icon: "success"
+                    });
+
+                    // Remove the deleted procurement from state
+                    setProcurements(prevProcurements => prevProcurements.filter(procurement => procurement.id !== procurementId));
+                }
             }
         } catch (error) {
             console.error('Greška prilikom brisanja nabavke:', error);
@@ -86,7 +102,7 @@ const ProcurementsComponent = ({auth, cosmetics}) => {
     return (
         <>
             <button onClick={()=>openModal(null)} className="bg-green-500 w-full hover:bg-green-700 text-white font-bold py-2 px-4 mb-3 rounded-10">Dodaj</button>
-            {isModalOpen && 
+            {isModalOpen &&
                 <ProcurementsFormModal
                     auth={auth}
                     closeModal={closeModal}
@@ -148,22 +164,19 @@ const ProcurementsComponent = ({auth, cosmetics}) => {
                                         >
                                             Ažuriraj
                                         </button>
-                                        <button 
+                                        <button
                                             className="bg-red-500 mb-2 w-24 hover:bg-red-300 text-white font-bold py-1 px-2 rounded"
-                                            onClick={() => {
-                                                if (window.confirm("Da li ste sigurni da želite da obrišete nabavku?")) {
-                                                    deleteProcurement(procurement.id);
-                                                }
+                                            onClick={() => {deleteProcurement(procurement.id);
                                             }}
                                         >
                                             Obriši
                                         </button>
-                                        
+
                                     </td>
                                 </tr>
                             );
                         })
-                    ) : 
+                    ) :
                     (
                         <tr>
                             <td className="md:px-6 lg:px-6 xl:px-6 2xl:px-6 py-4 whitespace-nowrap text-sm text-center font-medium text-gray-800 border-r">
